pyproject.toml
[project]
name = "deal-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "openai>=1.79.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
